+++ cmd/web_graceful.go
@@ -11,6 +11,7 @@ import (
 	"net/http"
 
 	"code.gitea.io/gitea/modules/log"
+
 	"github.com/facebookgo/grace/gracehttp"
 )