+++ models/migrations/migrations.go
@@ -102,6 +102,12 @@ var migrations = []Migration{
 	NewMigration("add show field in user openid table", addUserOpenIDShow),
 	// v26 -> v27
 	NewMigration("generate and migrate repo and wiki Git hooks", generateAndMigrateGitHookChains),
+	// v27 -> v28
+	NewMigration("change mirror interval from hours to time.Duration", convertIntervalToDuration),
+	// v28 -> v29
+	NewMigration("add field for repo size", addRepoSize),
+	// v29 -> v30
+	NewMigration("add commit status table", addCommitStatus),
 }
 
 // Migrate database to current version