+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * BREAKING
   * Password reset URL changed from `/user/forget_password` to `/user/forgot_password`
+  * SSH keys management URL changed from `/user/settings/ssh` to `/user/settings/keys`
+
 
 ## [1.1.0](https://github.com/go-gitea/gitea/releases/tag/v1.1.0) - 2017-03-09